requirements.txt
Django==4.2.1
psycopg2-binary==2.9.6
djangorestframework==3.14.0
requests==2.28.1
django-extensions==3.2.1
pyshorteners==1.0.1
factory_boy==3.2.1